Version 0.1.0-a0.2.0 Update
Introducing a rudimentary form of downloading the linked E2 file.
This linked E2 file is what the kernel will update automatically, whenever new updates are released.

Currently, its contents are printed to the chat as a place holder, because that data is yet to be saved to a local directory. Watch this space.